Everything now seems to work with SML/NJ 110.72.
The expansions from structure sharing to "where type" declarations are pretty ugly in places.  Figuring out all the required declarations for the functor applications in Elaboration.sml was not fun (think Apocalypse Now "The horror... the horror...").

It would be good to tidy things up now while it's fresh in my memory - now that the whole system compiles this can now be done incrementally.  On the other hand I may wait until my dreams stop being full of sharing specifications.  :-)