Tidied up some of the sharing and "where type" constraints.
There's still more that could be improved.  But, it's clear that the SML/NJ restrictions regarding definitional specs lead to a need to replace structure sharing with a multitude of explicit tedious type sharing/with-type specifications.  So it seems reasonable to leave as much as possible of this work for now - and consider alternative approaches such as following the style used in Twelf.