Fixed issue #4 by keeping the sort instances in constructor pattern-sorts so that when it "punts" on difficult instantiations, it uses the original sort for the constructor rather than inappropriately using the default sort.
It won't crash anymore, but the sorts assigned to pattern variables when it punts are still less accurate than they could be.  Next I'll be implementing a warning when this happens, and add a "try and see" approach with a configurable limit to determine which instantiations are feasible.

Also fixed an embarrassing tyvar ordering error in the last commit.  Sort instantiators need to be in the order of their *last* occurrences in the sort scheme.  This is more natural than first occurrences - particularly when applying constructors since then the order matches the order of the datatype parameters.